# api/openapi.yaml
openapi: 3.0.0
info:
  title: Phone Management API
  description: API for managing phones, users, accessories, and companies
  version: 1.0.0
  termsOfService: https://github.com/INF23DW1G25/REPORT_DW1_M2
  contact:
    name: M2 - G23
    email: random_email@umaia.pt
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: http://localhost:3000
    description: My localhost
    variables:
      basePath:
        default: /
paths:
...
  /company/{company_id}:
    get:
      tags:
        - Company
      operationId: getCompanyById
      security:
        - BearerAuth: []
      parameters:
        - name: company_id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: A company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
      x-swagger-router-controller: Company
    put:
      tags:
        - Company
      operationId: companyPUT
      security:
        - BearerAuth: []
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Company successfully updated
      x-swagger-router-controller: Company
    delete:
      tags:
        - Company
      operationId: companyCompany_idDELETE
      security:
        - BearerAuth: []
      parameters:
        - name: company_id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Company successfully deleted
      x-swagger-router-controller: Company
...
  /accessories/{phoneId}:
    get:
      tags:
        - Accessories
      description: Retrieve accessories by phone ID
      operationId: getAccessoriesByPhoneId
      security:
        - BearerAuth: []
      parameters:
        - name: phoneId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accessories'
        '404':
          description: Accessories not found
        '500':
          description: Internal server error
      x-swagger-router-controller: Accessories
    put:
      tags:
        - Accessories
      description: Update accessories by phone ID
      operationId: updateAccessoriesByPhoneId
      security:
        - BearerAuth: []
      parameters:
        - name: phoneId
          in: path
          required: true
          description: ID of the phone whose accessories to update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                screen_protector:
                  type: boolean
                case_protector:
                  type: boolean
      responses:
        '200':
          description: Accessories successfully updated
        '404':
          description: Phone not found
        '500':
          description: Internal server error
      x-swagger-router-controller: Accessories
components:
  securitySchemes:
    BearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Company:
      type: object
      properties:
        company_id:
          type: integer
          format: int64
        name:
          type: string
      example:
        company_id: 1
        name: Example Company
    Accessories:
      type: object
      properties:
        phone_id:
          type: integer
          format: int64
        screen_protector:
          type: boolean
        case_protector:
          type: boolean